Simple 16-bit program that runs on top of BIOS.
This program is going to evolve into a bootloader for the operating
system.

The program is written manually and has a simple instruction:
```
e9 fd ff
```
which means `jmp -3`.  It jumps into the beginning of the jump instruction.

The last two bytes of the 512-byte long program are the magic values
that indicate that the partition is bootable.  The program gets loaded
at the address 0x7c00 by the BIOS.

Bochs is a simple way to test this program.  The included
configuration files places the program on the bootable floppy disk.