Set up CMake and NASM assembler.
This commit sets up the CMake build system for programmatically
producting `boot-sect.bin` from a NASM source file.

The NASM source file performs the same function, however, it is
encoded in x86 differently.  NASM generates `eb fe` instead of `e9 f9
ff`.  The difference according to
[x86asm.net](http://ref.x86asm.net/geek.html#xEB) is in the size of
the operand: byte vs word-or-double-word.  The offest is also -2
instead of -3 since the command is longer.

CMake has a rudimentary and not very well documented support for
NASM.  It has even less support for generating 16-bit raw code.
Nevertheless, it can be made work.

The current way to run the program is:
```
$ cmake -S . -B build
$ make -C build run
```