Set up a simple C program and the necessary toolchain.
The C program simply returns a value that I have observed in the EAX
register.  It is a foundation for building the operating system
futher.

A cross-compiler is needed to build the C program for this operating
system.  The whole toolchain abstracts the architecture of the
target operating system from the architecture of the host computer.
In practice this amounts to building a GCC and linker tools with the
corresponding target parameters and using those binaries for
compilation.

I decided to take up CMake as part of this project.  It didn't make it
straightforward to make an obscure use-case like this work.

CMake is used for creating the toolchain as well as building the
operating system itself.  For example, one could perform these two
builds:

```
$ cmake -S . -B toolchain/ -Dbuild_the_toolchain=ON
$ make -j8 -C toolchain/
```

to build the toolchain and then,

```
$ cmake -S . -B build/
$ make -j8 run -C build/
```

to build the operating system using the toolchain.  Additional
configuration is needed to pass the right compiler flags for this kind
of free-standing low-level code as well as for calling the linker
directly.  The binary code for the C program can simply be appended to
the raw boot-sector binary.

The boot sector needs to be changed to call the right BIOS interrupts
for reading a few disk sectors where the C program is located.  The C
program gets loaded into the 1000h address, luckily it was linked with
such an assumption.  After reading the additional disk sectors, the
boot sectors calls "call 0x1000" to enter the world of C.