Compile 100% of the dependencies to reduce dependence on the host.
It's not always that you are able to have a convenient package
manager.  In particular, homebrew doesn't satisfy security
requirements of my work computer and therefore doesn't work.

This change instructs CMake to compile more of the GCC dependencies
and it eliminates the need for those dependencies to be installed on
the host computer.

The small change to kernel.c is just so that the program returns "CAFE"
in eax instead of "FECA" (no longer having "CAFE" in binary though,
because endianness).