Switch to the 32-bit protected mode.
This is done by configuring the Global Descriptor Table and setting
the control register in NASM.  I tried to arrange for it to happen in
the C layer, but there is no way to produce 16-bit code from GCC.  I
verified that its 32-bit output is not backward compatible even with
additional flags such as `-m16`.

The GDT sets up two simple segments one for the code and one for the data.
I verified that after this set up, the normal 32-bit code generated
by GCC works correctly.