Scale up the linking structure and set a few IDT handlers.
The structure of the operating system is now clearer.  A boot sector program is assembled and a kernel executable is assembled and the both are concatenated together.  The kernel executable is now assembled almost like a normal C binary where I can have libraries and dependencies.  That required being able to run NASM in `-f bin` format for the boot sector program but in `-f elf` format for the NASM libraries.

The two IDT handlers are set.  It is the next step although the support here is rudementary more so to test out the C infrastructure.