Fix the problem with linking and IDT handlers.
There was a problem with virtual addresses that prevent me from making further
progress with the IDT handlers.  I'd get "00014708026e[CPU0  ] interrupt():
gate descriptor is not valid sys seg (vector=0x01)" when executing an
interrupt.  Upon investigation it seemed like the addresses that GCC
assigns to the `idt_entries` array is at around 128th mb.  That's
problematic because this Bochs has 32mb RAM.

I got convinced that I need paging but then I decided to create an osdev
forum thread: https://forum.osdev.org/viewtopic.php?f=1&t=38290.  The
responses made my question the assumptions and I realized that I'm setting
the linker arguments on the wrong target.  A .bin gets appended to the
bootloader and I produce the .bin out of .elf.  I was setting the linker
operations on the .bin and .elf conversion instead of on the linking of
.elf.