Enable A20 gate at the boot.
Enabling A20 allows to access memory past 1MB.  There are a few ways that aren't fully
reliable.  This is the simplest one.  As it happens Bochs appears to have
A20 enabled.

The 15h BIOS interrupt with AX=2104h attemps to enable A20.  If successful it clears
CF, otherwise it sets CF.
A20 needs to enabled for entering the long 64-bit mode.