Switch to 64-bit mode instead of 32-bit protected.
This adds the code that switches from real mode to the 64 bit mode.
Official AMD documentation suggests to do so by switch to the 32-bit
mode first, yet I haven't found information on the downsides of this
more direct approach.  The only one I'm aware of is that the CPU will
triple fault in the case when it had to reboot anyway.