Upgrade the toolchain + IDT for x86_64-elf target.
The toolchain needs to be rebuilt.  NASM now generates elf64 format
in order to link to C libraries sucessfully.

A big difference is that in x86_64 first few parameters are passed
using registers instead of the stack.  That simplified a few places in
particular passing `Registers` by pointers to `isr_handler`.

pushad/popad aren't available and so I maintained parity with 32-bit for now
by manually pushing the same registers.

IDT and `Registers` structures grew to accomodate 64-bit pointers.