Prevent CMake from adding -rdynamic on Linux+GDT
For some reason CMake hard codes adding -rdynamic to the GCC command
line on Linux.  It doesn't do so on Mac OS X.  The option isn't supported
so I'm patching up CMake.

Another change here is I re-read the AMD manual and realized that I
applied the recommendation to aling the GDT structure for performance
reasons to the wrong structure.  I applied it to the pointer structure,
but I believe it recommends align the GDT structure itself.