Make sure that all of the boot loader is loaded into the memory.
Bootloader is stored in an image that mirrors a floppy.  It is stored
on the outer ring of that floppy.  The outer ring holds 8k memory.
With this change I'm trying to make sure that all of the boot loader
is loaded.  Right now it occupies less than the 8k, but I'm aiming for
a setup that is stable as I add code to the boot loader and it
occupies less memory.  Right now the `eh_frame` section from the .elf
gets copied into .bin and is used as a "filler" - it is loaded
partially because it ovelaps the 8k boundary.  That's okay, I don't
think we need that section.

I'm also annotating the memory layout of the .bin file with the text
files.  As I add code and the sections shift due to the alignment,
there is a chance that another, a more important, section might start
overlapping the 8k boundary.  That shouldn't happen because that
section will be loaded partially.