Pad the bootloader to be 9216 bytes.
Now we are always reading 9216 bytes of the outer ring of the floppy.
Should the bootloader be the smaller amount of bytes, it is
explicitly padded with 0 to be of that size.

I verified this by padding with /dev/random and observing that the
last byte gets loaded into the right position in memory.

The disadvantage is that this adds dependency on `dd`.  Oh well, I
couldn't make this work on Windows when I once tried and most unixes
have `dd`.